# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mystic Weave",
    "description": "Persistent game state backend for a text-based RPG powered by a custom GPT. The GPT is the narrator; this API is the memory. d100 roll-under resolution with domain scores and competency tiers.",
    "version": "5.3.1"
  },
  "servers": [
    {
      "url": "https://mysticweave-production.up.railway.app",
      "description": "Production (Railway)"
    }
  ],
  "paths": {
...
    "/location/{location_id}": {
      "get": {
        "summary": "Get Location",
        "description": "Load location data. The GPT must call this before describing any place.\nReturns 404 if the location does not exist in the database.",
        "operationId": "get_location_location__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/location": {
      "post": {
        "summary": "Upsert Location",
        "description": "Create or update a location via UPSERT, and persist valid world-graph connections from the payload.",
        "operationId": "upsert_location_location_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationResponse"
                }
              }
            }
          },
          "200": {
            "description": "Location updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/location/{location_id}/connections": {
      "get": {
        "summary": "Get Connections",
        "description": "Return all valid movement options from the given location.\n\nThe GPT may only move the player to locations listed here.\nMovement to unlisted locations is not permitted.",
        "operationId": "get_connections_location__location_id__connections_get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/scene/{session_id}": {
      "get": {
        "summary": "Get Scene Context",
        "description": "Build and return compact scene context from persisted state and current location for narration input.",
        "operationId": "get_scene_context_scene__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionInfo": {
        "properties": {
          "to_id": {
            "type": "string",
            "title": "To Id"
          },
          "traversal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traversal"
          },
          "distance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance"
          }
        },
        "type": "object",
        "required": [
          "to_id"
        ],
        "title": "ConnectionInfo"
      },
      "ConnectionsResponse": {
        "properties": {
          "from_id": {
            "type": "string",
            "title": "From Id"
          },
          "connections": {
            "items": {
              "$ref": "#/components/schemas/ConnectionInfo"
            },
            "type": "array",
            "title": "Connections"
          }
        },
        "type": "object",
        "required": [
          "from_id",
          "connections"
        ],
        "title": "ConnectionsResponse"
      },
...
      "FatigueState": {
        "type": "string",
        "enum": [
          "rested",
          "tired",
          "fatigued",
          "exhausted"
        ],
        "title": "FatigueState"
      },
...
      "HP": {
        "properties": {
          "current": {
            "type": "integer",
            "title": "Current"
          },
          "max": {
            "type": "integer",
            "title": "Max",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "current"
        ],
        "title": "HP"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HungerState": {
        "type": "string",
        "enum": [
          "sated",
          "hungry",
          "starving"
        ],
        "title": "HungerState"
      },
      "HydrationState": {
        "type": "string",
        "enum": [
          "hydrated",
          "thirsty",
          "dehydrated"
        ],
        "title": "HydrationState"
      },
...
      "LoadState": {
        "type": "string",
        "enum": [
          "light",
          "normal",
          "burdened",
          "overloaded"
        ],
        "title": "LoadState"
      },
      "LocationData": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "unknown"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "connections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Connections"
          },
          "threat_level": {
            "type": "integer",
            "title": "Threat Level",
            "default": 0
          },
          "known_npcs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known Npcs"
          },
          "discovered": {
            "type": "boolean",
            "title": "Discovered",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "LocationData",
        "description": "Shape of the JSONB data column in the locations table."
      },
      "LocationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "data": {
            "$ref": "#/components/schemas/LocationData"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "data"
        ],
        "title": "LocationResponse"
      },
...
      "SceneContext": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "location_summary": {
            "$ref": "#/components/schemas/SceneLocationSummary"
          },
          "visible_entities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Visible Entities"
          },
          "immediate_stakes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Immediate Stakes"
          },
          "relevant_character_state": {
            "$ref": "#/components/schemas/SceneRelevantCharacterState"
          },
          "recent_log": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recent Log"
          },
          "active_threats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Threats"
          },
          "available_opportunities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Opportunities"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "location_summary",
          "relevant_character_state"
        ],
        "title": "SceneContext",
        "description": "Derived GPT-facing scene payload assembled from full state + location + log."
      },
      "SceneLocationSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "unknown"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "SceneLocationSummary",
        "description": "Compact location payload for turn narration."
      },
...
      "SceneRelevantCharacterState": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "hp": {
            "$ref": "#/components/schemas/HP"
          },
          "status_effects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Status Effects"
          },
          "survival": {
            "$ref": "#/components/schemas/SurvivalState"
          },
          "active_companions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Companions"
          }
        },
        "type": "object",
        "required": [
          "name",
          "hp"
        ],
        "title": "SceneRelevantCharacterState",
        "description": "Narration-relevant subset of character/world state."
      },
...
      "SurvivalState": {
        "properties": {
          "hunger": {
            "$ref": "#/components/schemas/HungerState",
            "default": "sated"
          },
          "hydration": {
            "$ref": "#/components/schemas/HydrationState",
            "default": "hydrated"
          },
          "fatigue": {
            "$ref": "#/components/schemas/FatigueState",
            "default": "rested"
          },
          "load": {
            "$ref": "#/components/schemas/LoadState",
            "default": "normal"
          }
        },
        "type": "object",
        "title": "SurvivalState",
        "description": "Lightweight survival/load tracking using coarse state bands."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}